requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
cachetools==5.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-cors-headers==3.13.0
django-filter==21.1
django-import-export==2.8.0
djangorestframework==3.13.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
frozenlist==1.3.0
idna==3.3
inflection==0.5.1
ipinfo==4.2.0
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
multidict==6.0.2
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
